+++ src/components/__tests__/CvssCalculatorOverlayed.spec.ts
@@ -4,11 +4,11 @@ import sampleFlawFull from '@test-fixtures/sampleFlawFull.json';
 import { VueWrapper, mount } from '@vue/test-utils';
 import { describe, it, expect } from 'vitest';
 
-import { useFlawCvssScores, validateCvssVector } from '@/composables/useFlawCvssScores';
+import { useCvssScores, validateCvssVector } from '@/composables/useCvssScores';
 
 import { importActual } from '@/__tests__/helpers';
 
-vi.mock('@/composables/useFlawCvssScores');
+vi.mock('@/composables/useCvssScores');
 
 async function activateFactorButton(subject: any, factor: string, value: string) {
   const factorButtonGroup = subject.findAll('.btn-group-vertical')
@@ -30,7 +30,7 @@ describe('cvssCalculatorOverlayed', () => {
   let subject: Subject;
   beforeEach(async () => {
     if (subject) subject.unmount();
-    const { validateCvssVector: _validateCvssVector } = await importActual('@/composables/useFlawCvssScores');
+    const { validateCvssVector: _validateCvssVector } = await importActual('@/composables/useCvssScores');
 
     const cvssVector = ref<null | string>('');
     const cvssScore = ref<null | number>(0);
@@ -47,13 +47,13 @@ describe('cvssCalculatorOverlayed', () => {
     function updateScore(value: null | number) {
       cvssScore.value = value;
     }
-    vi.mocked(useFlawCvssScores).mockReturnValue({
+    vi.mocked(useCvssScores).mockReturnValue({
       cvssVector,
       cvssScore,
       cvssVersion,
       updateVector,
       updateScore,
-    } as ReturnType<typeof useFlawCvssScores>);
+    } as ReturnType<typeof useCvssScores>);
     vi.mocked(validateCvssVector).mockImplementation(_validateCvssVector);
 
     const importedComponent = await import('@/components/CvssCalculator/CvssCalculatorOverlayed.vue');
